refactor separate provider for opeani
